refactor: move searchAlgorithm to utils folder - Move searchAlgorithm to utils folder. - Add documentation to createRecipeCard method. - Remove a level of if/else inside checkAndApplyOptions function.
