+++ src/scripts/index.js
@@ -2,9 +2,9 @@ import { recipes } from '../data/recipes.js';
 import { Recipe } from './classes/Recipe.js';
 import { FilterList } from './classes/FilterList.js';
 import { enableDropdown } from './components/dropdownFilters.js';
-import { enableSelectFilter } from './components/searchAlgorithm.js';
-import { enableMainResearch } from './components/searchAlgorithm.js';
-import { showResultMessage } from './components/searchAlgorithm.js';
+import { enableSelectFilter } from './utils/searchAlgorithm.js';
+import { enableMainResearch } from './utils/searchAlgorithm.js';
+import { showResultMessage } from './utils/searchAlgorithm.js';
 
 // Initialize recipes page
 const initRecipesPage = () => {
